Refactor RecipientDetails
