+++ src/components/BankDetails/RecipientDetails/helpers/index.ts
@@ -2,3 +2,4 @@ export * from "./dot";
 export * from "./getDefaultValues";
 export * from "./isCountry";
 export * from "./isTextType";
+export * from "./mergeRequirements";
