+++ src/components/BankDetails/RecipientDetails/types.ts
@@ -1,6 +1,5 @@
-import { CreateRecipientRequest } from "types/aws";
-import { FileDropzoneAsset, OptionType } from "types/components";
-import { Country } from "types/components";
+import { AccountRequirements, CreateRecipientRequest } from "types/aws";
+import { Country, FileDropzoneAsset, OptionType } from "types/components";
 
 export type FormValues = Omit<
   CreateRecipientRequest,
@@ -15,4 +14,16 @@ export type FormValues = Omit<
 > & {
   bankStatementFile: FileDropzoneAsset;
   requirements: Record<string, null | string | OptionType<string> | Country>;
+};
+
+export type RequirementsData = {
+  active: boolean;
+  accountRequirements: AccountRequirements;
+  currentFormValues: FormValues;
+  refreshedRequirementsAdded: boolean;
+  /**
+   * Indicates whether requirements refresh is necessary.
+   * See https://docs.wise.com/api-docs/api-reference/recipient#account-requirements
+   */
+  refreshRequired: boolean;
 };
